Move original game to an inner module
