+++ pycross/app.py
@@ -220,32 +220,3 @@ class PycrossApp(App):
         if self.game_state.solved:
             self.query_one(GameGrid).disabled = True
             self.query_one(WinnerMessage).show()
-
-
-if __name__ == "__main__":
-    top_guides: Tuple[Tuple[int]] = (
-        (1,),
-        (1, 1, 1),
-        (2, 1),
-        (1, 1, 1),
-        (1,),
-    )
-
-    left_guides: Tuple[Tuple[int]] = (
-        (3,),
-        (1,),
-        (1, 1),
-        (1, 1, 1),
-        (1, 1),
-    )
-
-    solution: Tuple[Tuple[int]] = (
-        (0, 1, 1, 1, 0),
-        (0, 0, 1, 0, 0),
-        (0, 1, 0, 1, 0),
-        (1, 0, 1, 0, 1),
-        (0, 1, 0, 1, 0),
-    )
-
-    app = PycrossApp(top_guides, left_guides, solution)
-    app.run()
