+++ pycross/console_mode/game.py
